requirements.txt
algorithms==0.1.4
altgraph==0.17.3
ansicon==1.89.0
anyio==3.6.2
appdirs==1.4.4
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
astroid==2.12.13
asttokens==2.1.0
atomicwrites==1.4.0
attrs==21.4.0
auto-py-to-exe==2.23.1
auto1==0.0.1
autopep8==2.0.0
backcall==0.2.0
bcrypt==4.0.1
beautifulsoup4==4.11.1
black-box==1.0.2
bleach==5.0.1
blessed==1.19.1
bokeh==2.4.3
boltons @ file:///C:/b/abs_707eo7c09t/croot/boltons_1677628723117/work
bottle==0.12.23
bottle-websocket==0.2.9
brotlipy==0.7.0
caugetch==0.0.1
certifi==2022.12.7
cffi @ file:///C:/b/abs_49n3v2hyhr/croot/cffi_1670423218144/work
chardet==3.0.4
charset-normalizer==2.1.1
click==8.1.3
clipboard==0.0.4
cloudpickle==2.2.0
cmake==3.24.2
colorama==0.4.5
comtypes==1.1.14
conda==23.3.1
conda-content-trust @ file:///C:/Windows/TEMP/abs_4589313d-fc62-4ccc-81c0-b801b4449e833j1ajrwu/croots/recipe/conda-content-trust_1658126379362/work
conda-package-handling @ file:///C:/b/abs_fcga8w0uem/croot/conda-package-handling_1672865024290/work
conda_package_streaming @ file:///C:/b/abs_0e5n5hdal3/croot/conda-package-streaming_1670508162902/work
contourpy==1.0.6
cryptography @ file:///C:/b/abs_8ecplyc3n2/croot/cryptography_1677533105000/work
cycler==0.11.0
dask==2022.10.0
dask-sql==2022.10.1
debugpy==1.6.3
decorator==5.1.1
deepdiff==6.2.1
defusedxml==0.7.1
deprecation==2.1.0
dill==0.3.6
distlib==0.3.6
distributed==2022.10.0
dnspython==2.2.1
docopt==0.6.2
easyocr==1.6.2
ecdsa==0.18.0
Eel==0.14.0
email-validator==1.3.0
entrypoints==0.4
et-xmlfile==1.1.0
executing==1.2.0
fastapi==0.86.0
fastapi-jwt-auth==0.5.0
fastapi-sessions==0.3.2
fastjsonschema==2.16.2
filelock==3.8.0
fonttools==4.38.0
fpdf==1.7.2
fsspec==2022.10.0
future==0.18.3
getpass4==0.0.14.1
gevent==22.10.2
gevent-websocket==0.10.1
greenlet==2.0.0.post0
gTTS==2.2.4
h11==0.14.0
h2==3.2.0
HeapDict==1.0.1
hpack==3.0.0
hstspreload==2022.11.1
httpcore==0.16.1
httptools==0.5.0
httpx==0.23.1
hyperframe==5.2.0
idna==2.10
imageio==2.22.4
img2pdf==0.4.4
importlib-metadata==6.0.0
inflect==6.0.2
iniconfig==1.1.1
inquirer==3.1.2
ipykernel==6.17.0
ipython==8.6.0
ipython-genutils==0.2.0
ipywidgets==8.0.2
isort==5.10.1
itsdangerous==2.1.2
jedi==0.18.1
Jinja2==3.1.2
jinxed==1.2.0
joblib==1.2.0
jsonpatch @ file:///tmp/build/80754af9/jsonpatch_1615747632069/work
jsonpointer==2.1
jsonschema==4.17.0
jupyter-client==7.3.4
jupyter-server==1.21.0
jupyter_core==4.11.2
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.3
kiwisolver==1.4.4
lazy-object-proxy==1.8.0
lazy_loader==0.2
locket==1.0.0
lxml==4.9.1
MarkupSafe==2.1.1
matplotlib==3.6.2
matplotlib-inline==0.1.6
mccabe==0.7.0
menuinst @ file:///C:/Users/BUILDE~1/AppData/Local/Temp/abs_455sf5o0ct/croots/recipe/menuinst_1661805970842/work
mistune==2.0.4
MouseInfo==0.1.3
msgpack==1.0.4
mypy==1.1.1
mypy-extensions==1.0.0
nbclassic==0.4.8
nbclient==0.7.0
nbconvert==7.2.3
nbformat==5.7.0
nest-asyncio==1.5.6
networkx==2.8.8
ninja==1.11.1
nltk==3.7
notebook==6.5.2
notebook_shim==0.2.2
numpy==1.23.4
omymodels==0.13.0
opencv-contrib-python==4.6.0.66
opencv-python==4.6.0.66
opencv-python-headless==4.5.4.60
openpyxl==3.0.10
ordered-set==4.1.0
orjson==3.8.3
packaging==21.3
pandas==1.5.2
pandocfilters==1.5.0
parsimonious==0.8.1
parso==0.8.3
partd==1.3.0
passlib==1.7.4
pdf2image==1.16.0
pdfminer==20191125
pefile==2022.5.30
pep8==1.7.1
pickleshare==0.7.5
pikepdf==6.2.8.post1
Pillow==9.3.0
pip==23.1.2
pipreqs==0.4.11
platformdirs==2.5.2
pluggy @ file:///C:/ci/pluggy_1648042746254/work
ply==3.11
polars==0.16.12
prometheus-client==0.15.0
prompt-toolkit==3.0.32
psutil==5.9.3
pure-eval==0.2.2
py==1.11.0
py-models-parser==0.6.0
py4j==0.10.9.5
pyarmor==7.7.4
pyarrow==10.0.1
pyasn1==0.4.8
PyAutoGUI==0.9.53
pyclipper==1.3.0.post4
pycodestyle==2.9.1
pycosat @ file:///C:/b/abs_4b1rrw8pn9/croot/pycosat_1666807711599/work
pycparser @ file:///tmp/build/80754af9/pycparser_1636541352034/work
pycryptodome==3.15.0
pycryptodomex==3.17
pydantic==1.10.2
PyGetWindow==0.0.9
Pygments==2.13.0
pyinstaller==5.6.2
pyinstaller-hooks-contrib==2022.13
PyJWT==1.7.1
pylint==2.15.6
pyminizip @ file:///D:/bld/pyminizip_1650029370958/work
pymongo==4.3.2
PyMsgBox==1.0.9
PyMuPDF==1.21.0
pyodbc==4.0.34
pyOpenSSL @ file:///C:/b/abs_552w85x1jz/croot/pyopenssl_1677607703691/work
pyparsing==3.0.9
pypdf==3.8.1
PyPDF2==2.11.1
pyperclip==1.8.2
pypiwin32==223
PyRect==0.2.0
pyrsistent==0.19.2
PyScreeze==0.1.28
PySocks @ file:///C:/ci_310/pysocks_1642089375450/work
pytesseract==0.3.10
pytest==7.1.2
python-bidi==0.4.2
python-dateutil==2.8.2
python-dotenv==0.21.0
python-editor==1.0.4
python-jose==3.3.0
python-multipart==0.0.5
pyttsx3==2.90
pytweening==1.0.4
pytz==2022.6
pytz-deprecation-shim==0.1.0.post0
PyWavelets==1.4.1
pywin32==306
pywin32-ctypes==0.2.0
pywinpty==2.0.10
PyYAML==6.0
pyzmq==24.0.1
qtconsole==5.4.0
QtPy==2.2.1
readchar==4.0.3
regex==2022.10.31
requests @ file:///C:/b/abs_80_edxtvhu/croot/requests_1678709752318/work
rfc3986==1.5.0
rsa==4.9
ruamel.yaml @ file:///C:/b/abs_30ee5qbthd/croot/ruamel.yaml_1666304562000/work
ruamel.yaml.clib @ file:///C:/b/abs_aarblxbilo/croot/ruamel.yaml.clib_1666302270884/work
scikit-image==0.20.0
scipy==1.9.3
Send2Trash==1.8.0
setuptools==67.7.2
Shapely==1.8.5.post1
simple-ddl-parser==0.28.1
six @ file:///tmp/build/80754af9/six_1644875935023/work
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
sqlacodegen==3.0.0b2
SQLAlchemy==1.4.41
sqlalchemy2-stubs==0.0.2a32
sqlmodel==0.0.8
stack-data==0.6.0
starlette==0.20.4
table-meta==0.1.5
tabulate==0.9.0
tblib==1.7.0
tenacity==8.1.0
terminado==0.17.0
threadpoolctl==3.1.0
tifffile==2022.10.10
tinycss2==1.2.1
tomli==2.0.1
tomlkit==0.11.6
toolz @ file:///C:/b/abs_cfvk6rc40d/croot/toolz_1667464080130/work
torch==1.13.0
torchaudio==0.13.0
torchvision==0.14.0
tornado==6.1
tqdm==4.64.1
traitlets==5.5.0
typing_extensions==4.4.0
tzdata==2022.6
tzlocal==4.2
ujson==5.6.0
urllib3==1.26.12
uvicorn==0.20.0
virtualenv==20.16.5
Wand==0.6.10
watchdog==3.0.0
watchfiles==0.18.1
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.4.2
websockets==10.4
wheel==0.38.4
whichcraft==0.6.1
widgetsnbextension==4.0.3
win-inet-pton @ file:///C:/ci_310/win_inet_pton_1642658466512/work
wincertstore==0.2
wrapt==1.14.1
xlrd==2.0.1
XlsxWriter==3.0.3
yarg==0.1.9
zict==2.2.0
zipp==3.10.0
zope.event==4.5.0
zope.interface==5.5.1
zstandard==0.19.0
